fix(LogsVolumePanel): refactor series visibility
