+++ src/services/panel.ts
@@ -1,10 +1,13 @@
 import { DataFrame, FieldConfig, FieldMatcherID } from '@grafana/data';
 import {
+  FieldConfigBuilder,
   FieldConfigBuilders,
   FieldConfigOverridesBuilder,
   PanelBuilders,
   SceneDataTransformer,
+  SceneObject,
   SceneQueryRunner,
+  VizPanel,
 } from '@grafana/scenes';
 import { map, Observable } from 'rxjs';
 import { LokiQuery } from './query';
@@ -12,6 +15,7 @@ import { HideSeriesConfig } from '@grafana/schema';
 import { WRAPPED_LOKI_DS_UID } from './datasource';
 import { LogsSceneQueryRunner } from './LogsSceneQueryRunner';
 import { DrawStyle, StackingMode } from '@grafana/ui';
+import { getLabelsFromSeries, getVisibleLevels } from './levels';
 
 const UNKNOWN_LEVEL_LOGS = 'logs';
 export function setLevelColorOverrides(overrides: FieldConfigOverridesBuilder<FieldConfig>) {
@@ -74,6 +78,18 @@ export function setLevelSeriesOverrides(levels: string[], overrideConfig: FieldC
   const overrides = overrideConfig.build();
   // @ts-expect-error
   overrides[overrides.length - 1].__systemRef = 'hideSeriesFrom';
+}
+
+export function syncLogsPanelVisibleSeries(panel: VizPanel, series: DataFrame[], sceneRef: SceneObject) {
+  const focusedLevels = getVisibleLevels(getLabelsFromSeries(series), sceneRef);
+  if (focusedLevels?.length) {
+    const config = setLogsVolumeFieldConfigs(FieldConfigBuilders.timeseries()).setOverrides(
+      setLevelSeriesOverrides.bind(null, focusedLevels)
+    );
+    if (config instanceof FieldConfigBuilder) {
+      panel.onFieldConfigChange(config.build(), true);
+    }
+  }
 }
 
 export function sortLevelTransformation() {
